Initial Upload
The script works but is coded with an example of my system and how I use it. If you use this script you need to change your destination_directory accordingly. You may also want to put the output log somewhere other than the folder the script is run from.

To Do:
Create a config file to set the parameters from plain text.
Create an exclusions list to automatically add and failed files/directories to the list.